refactor: decompose ACCanvas.tsx into focused component hierarchy (Steps 4–8)
ACCanvas.tsx reduced from ~2175 lines to 298 lines. Extracted all logic into
dedicated modules: shared UI components, modals, search components, view
components, and custom hooks. Pure structural refactoring — no visual or
behavioral changes.
